Rename conditional_accept to detect_sni
