+++ src/transport/tls/mod.rs
@@ -1,6 +1,6 @@
 pub mod client;
-mod conditional_accept;
 mod connection;
+mod detect_sni;
 mod io;
 pub mod listen;
